==> whisky_web-scarping/3_link_cleaning.ipynb <==
import pandas as pd 

top100_whisky = pd.read_csv('Top100_whisky.csv')
for col in top100_whisky.columns[2:]:
    top100_whisky[col] = top100_whisky[col].apply(lambda x : x.lstrip('[').rstrip(']'))
top100_whisky.head()

top100_whisky = pd.read_csv('Top100_whisky.csv')

bottle = 4
for info in top100_whisky.iloc[bottle]:
    print (info, '\n')

# hkls  ========update by name 
top100_whisky.iloc[bottle,2] = None
top100_whisky.iloc[bottle,3] = None

# hkls  ========update by select one item from list 
name = top100_whisky.iloc[bottle,2].split(',')
link = top100_whisky.iloc[bottle,3].split(',')
top100_whisky.iloc[bottle,2] = name[1]
top100_whisky.iloc[bottle,3] = link[1]

# price  ========update by name 
top100_whisky.iloc[bottle,4] = None
top100_whisky.iloc[bottle,5] = None

# price  ========update by select one item from list 
name = top100_whisky.iloc[bottle,4].split(',')
link = top100_whisky.iloc[bottle,5].split(',')
top100_whisky.iloc[bottle,4] = name[1]
top100_whisky.iloc[bottle,5] = link[1]

# watsons ========update by name 
top100_whisky.iloc[bottle,6] = ''
top100_whisky.iloc[bottle,7] = ''

for info in top100_whisky.iloc[bottle]:
    print (info, '\n')

top100_whisky.to_csv('Top100_whisky.csv', index=False)

# To be updated: 


# Cannot find in HK
# 17: Glenfarclas 15
# 35: Edradour 10 
# 44: Bakers 7
# 45: Edradour 12
# 73: Blair Athol 12 
# 90: Redbreast 15


# Remark 
## 55: Glen Garioch 12, Single Malt Whisky(hkls), Virgin Oak (price.com)
## 59: Benriach 10,  Take -Single Malt Scotch Whisky, give up The BenRiach Curiositas 10years(picture is the white bottle)
## 60: Benriach 15,  Take -Pedro Ximenez Finish Single Malt Whisky/ give up Wood Finish Single Malt Whisky
## 41: Macallan 12,  Take -Sherry Oak 
## 89: Macallan 18,  Take -Sherry Oak 

top100_whisky = pd.read_csv('Top100_whisky.csv')

top100_whisky.info()

for r in range (len(top100_whisky['Name'])):
    for c in range(2,8):
        try:
            top100_whisky.iloc[r,c] = top100_whisky.iloc[r,c].lstrip(" '").rstrip(" '")
        except:
            pass

top100_whisky.head()

top100_whisky.to_csv('Top100_whisky.csv', index=False)
