==> whisky_web-scarping/4_webscraping_hkls_link.ipynb <==
import pandas as pd
from bs4 import BeautifulSoup 
import requests 
import re

top100_whisky = pd.read_csv('Top100_whisky.csv')
top100_whisky.head()

## Create dataframe 

# create whisky details csv for hkls 
hkls_details = top100_whisky.iloc[:,0:4].copy()
hkls_details.head()

# add new columns in hkls_details
hkls_details['Price'] = pd.Series(dtype='str')
hkls_details['Stock'] = pd.Series(dtype='str')
hkls_details['Volume'] = pd.Series(dtype='str')

for w in range(len(hkls_details['Name'])):
    if len(str(hkls_details['hkls_name'][w]))>5:
        URL = hkls_details['hkls_link'][w]
        html = requests.get(URL)
        page = BeautifulSoup(html.text,'html.parser')

        hkls_details['Price'][w] = page.find(class_ = 'price').text
        if page.find (class_ = 'stock available'):
            hkls_details['Stock'][w] = True
        else:
            hkls_details['Stock'][w] = False
        try:
            hkls_details['Volume'][w] = re.search(r'(Alcohol)(\s)(\d+\w+)$', page.find(class_= 'product attribute overview').text)[3]
        except:
            hkls_details['Volume'][w] = 'Not available'

hkls_details.to_csv('hkls_details.csv', index = False)

## Update dataframe 

def update_hkls():
    import pandas as pd
    from bs4 import BeautifulSoup 
    import requests 
    import re

    hkls_details = pd.read_csv('hkls_details.csv')

    # update Price and Stock
    for w in range(len(hkls_details['Name'])):
        if len(str(hkls_details['hkls_name'][w]))>5:
            URL = hkls_details['hkls_link'][w]
            html = requests.get(URL)
            page = BeautifulSoup(html.text,'html.parser')

            hkls_details['Price'][w] = int(re.search(r'\$\d+,\d+|\$\d+',page.find(class_ = 'price').text).group().replace('$', '').replace(',',''))
            if page.find (class_ = 'stock available'):
                hkls_details['Stock'][w] = True
            else:
                hkls_details['Stock'][w] = False
            
    hkls_details.to_csv('hkls_details.csv', index = False)
    return hkls_details


update_hkls()

hkls_details.head()

## Select whisky

def hkls_whisky(index):
    hkls_details = pd.read_csv('hkls_details.csv')
    name, address, price = None, None, None

    if hkls_details['Stock'][index] == True:
        name = 'Hong Kong Liquor Store (10 Stores with online shop)'
        address = 'https://www.hkliquorstore.com/locations'
        price = hkls_details['Price'][index]

    return [name,address,price]

# hkls_details['Stock'][12] ---> no stock 
hkls_whisky(12)

# hkls_details['Stock'][14] ---> have stock 
hkls_whisky(14)

# hkls_details['Stock'][17] ---> not selling 
hkls_whisky(17)
