==> whisky_web-scarping/5_webscraping_wastons_link.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.by import By
import time
from selenium.webdriver.common.keys import Keys
from webdriver_manager.chrome import ChromeDriverManager

import pandas as pd 
import re 

# To close the web browser 
# from selenium import webdriver   # for webdriver
# from selenium.webdriver.support.ui import WebDriverWait  # for implicit and explict waits
# from selenium.webdriver.chrome.options import Options  # for suppressing the browser

## create dataframe

top100_whisky = pd.read_csv('Top100_whisky.csv')

# create whisky details csv for hkls 
whisky_name = top100_whisky.iloc[:,0:2].copy()
wastons_name = top100_whisky.iloc[:,6:8].copy()
watsons_details = pd.concat([whisky_name, wastons_name], axis = 1)
watsons_details.head()

# add new columns in hkls_details
watsons_details['Price'] = pd.Series(dtype='str')
watsons_details['Stock'] = pd.Series(dtype='str')
watsons_details['Volume'] = pd.Series(dtype='str')

# obtain links 
page = watsons_details['watsons_link']

# Web scrape price, volume and stock
for index, p in enumerate(page):
    if len(str(p))>5:
    # Oillio's code to stop the banner
        capa = DesiredCapabilities.CHROME
        capa["pageLoadStrategy"] = "none"
        driver = webdriver.Chrome('../chromedriver',desired_capabilities=capa)
        driver.get(p)
        time.sleep(5)
        driver.execute_script("window.stop();")
        driver.maximize_window()    

        try:
            watsons_details['Price'][index] = re.search(r'(\$)(\d+,\d+)|(\$)(\d+)', driver.find_element_by_class_name("offer-price").find_element_by_class_name("price").text).group(2)
        except:
            watsons_details['Price'][index]= re.search(r'(\$)(\d+,\d+)|(\$)(\d+)',driver.find_element_by_class_name("regular-price").text).group(2)
                    
        try: 
            if driver.find_element_by_class_name("addProduct").text == 'Out of Stock':
                watsons_details['Stock'][index] = False
            else:
                watsons_details['Stock'][index] = True
        except:
            watsons_details['Stock'][index] = True
        try:
            watsons_details['Volume'][index] = str(int(re.search(r'(\d+)(CL)$', driver.find_element_by_xpath('//*[@id="overview"]/div/div/div[2]/table/tbody/tr[1]/td[2]').text).group(1))*10)+'ml'
        except:
            watsons_details['Volume'][index] = 'Not Available'

watsons_details.to_csv('watsons_details.csv', index=False)

## Update dataframe 

def update_watsons_detail():
    watsons_details = pd.read_csv('watsons_details.csv')
    page = watsons_details['watsons_link']

    for index, p in enumerate(page):
        if len(str(p))>5:
        # Oillio's code to stop the banner
            capa = DesiredCapabilities.CHROME
            capa["pageLoadStrategy"] = "none"
            driver = webdriver.Chrome('../chromedriver',desired_capabilities=capa)
            driver.get(p)
            time.sleep(7)
            driver.execute_script("window.stop();")
            driver.maximize_window()    

            try:
                watsons_details['Price'][index]= int(re.search(r'\$\d+,\d+|\$\d+',driver.find_element_by_class_name("regular-price").text).group().replace(',', '').replace('$', ''))
            except:
                watsons_details['Price'][index] = int(re.search(r'\$\d+,\d+|\$\d+', driver.find_element_by_class_name("offer-price").find_element_by_class_name("price").text).group().replace(',', '').replace('$', ''))
            try: 
                if driver.find_element_by_class_name("addProduct").text == 'Out of Stock':
                    watsons_details['Stock'][index] = False
                else:
                    watsons_details['Stock'][index] = True
            except:
                watsons_details['Stock'][index] = True
      
    watsons_details.to_csv('watsons_details.csv', index=False)
    return watsons_details

update_watsons_detail()
    

watsons_details=pd.read_csv('watsons_details.csv')
watsons_details.head()

## Select Whisky  

def watsons_whisky(index):
    watsons_details = pd.read_csv('watsons_details.csv')
    name, address, price = None, None, None

    if watsons_details['Stock'][index] == True:
        name = 'Watson\'s wine  (23 Stores with online shop)'
        address = 'https://www.watsonswine.com/en'
        price = watsons_details['Price'][index]

    return [name,address,price]

# watsons_whisky['Stock'][2] ---> with stock
watsons_whisky(2)

# watsons_whisky['Stock'][3] ---> no stock
watsons_whisky(3)

# watsons_whisky['Stock'][4] ---> not selling
watsons_whisky(4)
